Modify update computation to reuse already computed values.
By rewriting the equations it is possible to introduce the same terms in
the for monomers and polymers. As the term has already been computed for
the polymers using it in monomers introduces no cost.

Numerically, it has been observed that this change gives 10% speedup for
the CPU version running initialized_0005 config.