Ensure that only valid particles are selected.
Due to how summation of small numbers work, it might have been the case 
that find procedure will iterate over the whole list of particles and 
still won't find a pair. We fix it but keeping track of all valid 
particles and assigning the last one in case the whole list is exhausted 
without success.