Fix the computation of simulation time by remembering increased in cell size.
The max_num_particles only counted the maximum number of particles throughout the evolution of the system.
However, the simulation time depends on the number of initial particles in the whole simulation, i.e.
total number of monomers. The latter is different from the former due to the former counting all clusters as single particle.